Rename apis
