--- src/api/index.ts
@@ -1,6 +1,0 @@
-export { getUserAlbums } from './album';
-export { getPostComments, getUserPostComments } from './comment';
-export { getAlbumPhotos, getAlbumPhoto } from './photo';
-export { getPosts, getPost, getUserPosts, getUserPost } from './post';
-export { getUserTodos } from './todo';
-export { getUser } from './user';
