+++ src/app/profile/albums/[albumId]/photos/page.tsx
@@ -1,7 +1,7 @@
 import clsx from 'clsx';
 import Image from 'next/image';
 import Link from 'next/link';
-import { getAlbumPhotos } from '@/api';
+import { getAlbumPhotos } from '@/apis';
 import type { Photo } from '@/types';
 
 /**
